V5.2-000

+++ sr_port/ionl_rdone.c
@@ -1,6 +1,6 @@
 /****************************************************************
  *								*
- *	Copyright 2001 Sanchez Computer Associates, Inc.	*
+ *	Copyright 2001, 2006 Fidelity Information Services, Inc	*
  *								*
  *	This source code contains the intellectual property	*
  *	of its copyright holder(s), and is made available	*
@@ -14,7 +14,7 @@
 #include "mdef.h"
 #include "io.h"
 
-short	ionl_rdone(mint *val, int4 timeout)
+int	ionl_rdone(mint *val, int4 timeout)
 {
 	mval	tmp;
