+++ sr_port/ionl_read.c
@@ -1,6 +1,6 @@
 /****************************************************************
  *								*
- *	Copyright 2001 Sanchez Computer Associates, Inc.	*
+ *	Copyright 2001, 2006 Fidelity Information Services, Inc	*
  *								*
  *	This source code contains the intellectual property	*
  *	of its copyright holder(s), and is made available	*
@@ -12,7 +12,7 @@
 #include "mdef.h"
 #include "io.h"
 
-short	ionl_read(mval *val, int4 timeout)
+int	ionl_read(mval *val, int4 timeout)
 {
 	return ionl_readfl(val, 1, timeout);
 }
