+++ sr_port/ionl_readfl.c
@@ -1,6 +1,6 @@
 /****************************************************************
  *								*
- *	Copyright 2001 Sanchez Computer Associates, Inc.	*
+ *	Copyright 2001, 2006 Fidelity Information Services, Inc	*
  *								*
  *	This source code contains the intellectual property	*
  *	of its copyright holder(s), and is made available	*
@@ -14,7 +14,7 @@
 
 GBLREF io_pair io_curr_device;
 
-short ionl_readfl(mval *val, int4 width, int4 timeout)
+int ionl_readfl(mval *val, int4 width, int4 timeout)
 {
 	io_desc *dv;
 	error_def(ERR_IOEOF);
