+++ sr_unix/iott_read.c
@@ -1,6 +1,6 @@
 /****************************************************************
  *								*
- *	Copyright 2001 Sanchez Computer Associates, Inc.	*
+ *	Copyright 2001, 2006 Fidelity Information Services, Inc	*
  *								*
  *	This source code contains the intellectual property	*
  *	of its copyright holder(s), and is made available	*
@@ -16,9 +16,8 @@
 
 GBLREF	io_pair		io_curr_device;
 
-short	iott_read (mval *v, int4 timeout)
-    	   
-    	        	/* timeout in seconds */
+int	iott_read (mval *v, int4 timeout)   /* timeout in seconds */
+
 {
 	d_tt_struct	*tt_ptr;
