+++ sr_unix/ious_rdone.c
@@ -1,6 +1,6 @@
 /****************************************************************
  *								*
- *	Copyright 2001 Sanchez Computer Associates, Inc.	*
+ *	Copyright 2001, 2006 Fidelity Information Services, Inc	*
  *								*
  *	This source code contains the intellectual property	*
  *	of its copyright holder(s), and is made available	*
@@ -15,7 +15,7 @@
 
 GBLREF io_pair		io_curr_device;
 
-short ious_rdone(mint *v, int4 t)
+int ious_rdone(mint *v, int4 t)
 {
 	*v = -1;
 	return TRUE;
