+++ sr_unix/ious_read.c
@@ -1,6 +1,6 @@
 /****************************************************************
  *								*
- *	Copyright 2001 Sanchez Computer Associates, Inc.	*
+ *	Copyright 2001, 2006 Fidelity Information Services, Inc	*
  *								*
  *	This source code contains the intellectual property	*
  *	of its copyright holder(s), and is made available	*
@@ -17,7 +17,7 @@
 GBLREF spdesc		stringpool;
 GBLREF io_pair		io_curr_device;
 
-short ious_read(mval *v, int4 t)
+int ious_read(mval *v, int4 t)
 {
 	v->str.len = 0;
 	v->str.addr = (char *)0;
